feat(create-pull-request.yml): add GitHub Actions workflow to create pull requests
This commit adds a new GitHub Actions workflow that creates a pull
request when a new branch is created. The workflow runs on the `create`
event and is triggered when a new branch is created. The workflow uses
the `repo-sync/pull-request` action to create the pull request. The pull
request is created from the newly created branch to the `develop`
branch. The title of the pull request is "Pulling [branch name] into
main", where [branch name] is the name of the newly created branch.