refactor(eslint.yaml): remove continue-on-error flag and add if statement to merge all eslint reports
The continue-on-error flag has been removed from the generate eslint
report step. Instead, an if statement has been added to the merge all
eslint reports step to only run if the previous step failed. This
ensures that the eslint report is only merged if there are errors in the
code.