chore(package.json): add --passWithNoTests flag to coverage script
The --passWithNoTests flag is added to the coverage script in all package.json files. This flag allows the coverage script to pass even if there are no tests to run. This is useful in cases where a package has no tests yet, but the coverage script is still run as part of the build process.