feat(package.json): add coverage script
This commit adds a new script to the package.json file called "coverage". This script runs the "turbo run coverage --continue" command, which generates a code coverage report for the project. The "--continue" option allows the task to run for all packages even if an error occurs in a task run on another package.