chore(dicty-editor): update package name and dependencies
The package name has been changed from "lexical-editor" to "dicty-editor" to better reflect the purpose of the package. The "dictybase-toolbar" dependency has been replaced with "editor-toolbar" to improve the functionality of the package. The "eslint-config-dictybase" dependency has been replaced with "eslint-config-custom" to allow for custom configuration of the linter. The "@tsconfig/dictybase" and "typescript" dependencies have been removed as they are no longer needed. The "tsconfig" dependency has been added to allow for custom configuration of the TypeScript compiler.