chore(dicty-editor): update tsconfig.json
The tsconfig.json file has been updated to extend the react-library-esm.json configuration file, which provides a base configuration for creating a React library that can be consumed as an ES module. The include and exclude properties have also been updated to include the current directory and exclude the dist, build, and node_modules directories.