refactor(Editor.tsx): change content prop to include an id field
The content prop has been updated to include an id field, which will be used to identify the content. The initialEditorState is initialized to null if the editorState property is falsy. This ensures that the editor will always have a valid initial state, even if the content prop is not provided.