refactor(useLocalStorage.ts): change useLocalStorage hook to named export functions
The useLocalStorage hook has been refactored to export two named functions, saveLocalStorage and retrieveLocalStorage. This change improves the reusability of the functions and makes them easier to test. The functions take in a LexicalEditor instance and a storageKey string as arguments. The saveLocalStorage function saves the editor state to local storage, while the retrieveLocalStorage function retrieves the editor state from local storage and sets it to the editor instance.