refactor(useButtonState.ts): remove unnecessary finally block
The finally block is not necessary in this case as it does not add any value to the code. Previously in the finally block, a setTimeout would be created that would change the state of the button after a certain interval. But since the entire button component unmounts after successful save, this would lead to an error. Thus, the timeout has been moved to the error block.