feat(Editor.tsx): add handleSaveClick prop to Editor component
The SaveButton component is imported from the persistence-plugin package, which has been moved to the Editor component. The Editor component now accepts a handleSaveClick prop, which is a function that is called when the SaveButton component is clicked. This allows the Editor component to (eventually) save its data to a back end, and well as initiated a page navigation.