refactor(LocalPersistencePlugin.tsx): remove unused code and simplify SAVE_LOCAL_COMMAND and RETRIEVE_LOCAL_COMMAND
The RETRIEVE_LOCAL_COMMAND command is simplified to only call retrieveLocalStorage with the storageKey property. The SAVE_LOCAL_COMMAND command is also simplified to not require a payload as the storageKey is now passed as a property to the LocalPersistencePlugin component. Since the retrieveLocalStorage function is only needed to be called once when the editor loads, its command listener has been removed.