feat(news-component): add news component package
The news component package is added to the project. It includes a NewsContainer component that displays a list of news articles and a NewsHeader component that displays the title of the news section and an icon for adding new articles. The package also includes a README file and a package.json file with dependencies and scripts. The NewsContainer component currently uses a sample list of articles, but will be updated to use a GraphQL query in the future.