feat(news-component): add BrowseNews component
The BrowseNews component is added to the news-component package. It renders a list of news articles and uses the useListNewsContentQuery hook from the dicty-graphql-schema package to fetch the data. The NewsPreview component is also added to the package, which is used to render each individual news article in the list. The index.ts file is updated to export the BrowseNews component.
issue #173