refactor(mockListNewsContent.ts): change slug values to be more descriptive
The slug values for each news item have been changed to be more descriptive by prefixing them with "mock-". This improves the readability of the code and makes it easier to identify each news item. Additionally, the updatedAt field has been updated to use the ISO string format for consistency.