feat(news-component): add DeleteButton component and use it in DisplayNews
The DeleteButton component is a new component that allows users to delete news articles. It is used in the DisplayNews component to provide a way for users to delete the news article they are currently viewing. The DisplayNews component has also been updated to use the DeleteButton component and to format the updatedAt date using the formatDateISOString utility function.
issue #173