refactor(AuthStore.tsx): extract mockUser object and add testPermission
The mockUser object is now extracted from the MockSuperuser object and includes a new permission, testPermission. This improves the readability of the code and makes it easier to modify the user object in the future. The testPermission is added to the superUserRole object, which is then added to the roles array of the mockUser object.