feat(news-component): add useCreateNews and useUpdateNews hooks
Two new hooks, useCreateNews and useUpdateNews, have been added to the news-component package. These hooks are used to create and update news content respectively. They both use the useCreateContentMutation and useUpdateContentMutation hooks from the dicty-graphql-schema package to make the necessary GraphQL mutations. The hooks also use the slugify package to generate a slug from the title of the news content. Finally, the hooks use the useNavigate hook from react-router-dom to navigate to the newly created or updated news content.