feat(dicty-frontpage): add database initialization and mock data
This commit adds a new file, index.ts, which initializes a new BrowserLevel database instance with the name "dicty-frontpage" and value encoding set to JSON. It also exports the database instance for use in other modules. The initializeData function is commented out, but it would batch insert the mock news content data into the database.
The database mock data is also updated in database.ts to use the new superuserProperties object instead of the deprecated MockSuperuser object. The superuserRole object is also used instead of the deprecated superUserRole object.