chore(dicty-editor): remove unnecessary files and configurations
The files and configurations removed are no longer needed in the project as it is no longer intended to be a standalone application. The index.html file and vite.svg file were deleted as they were not being used in the project. The tsconfig.node.json and vite.config.ts files were also deleted as they were not being used in the project and were unnecessary.