feat(dicty-graphql-schema): add limit argument to listContent query
The listContent query now accepts a limit argument, which allows the client to specify the number of items to return. This improves the flexibility of the query and reduces the amount of data returned to the client. The schema.json and content.graphql files were updated to reflect this change. The useListNewsContentQuery function was also updated to require the limit argument.