feat(BrowseNews.tsx): add pagination to news list
This commit adds pagination to the news list by using the `Pagination` component. The `articlesInRange` atom is used to slice the `listContent` array returned by the `useListNewsContentQuery` hook. The `articlesListTotalAtom` atom is used to set the total number of articles in the list. This allows the `Pagination` component to calculate the number of pages needed to display all the articles.
issue #231