feat(dicty-frontpage): add VITE_APP_SEED_MOCK environment variable and dev:seed script
The VITE_APP_SEED_MOCK environment variable is added to the .env.seed file, which is used to seed the application with mock data. The dev:seed script is added to the package.json file, which starts the application in seed mode. This allows developers to test the application with mock data during development.