feat(Links.tsx): add Links component to render FooterLinks
The Links component is added to render the FooterLinks component. It takes in an array of FooterItem objects and maps them to FooterLink components. The resulting array of components is then passed to the FooterLinks component. If the array is empty, an error message is displayed. The component uses functional programming concepts from fp-ts library.
issue #250