feat(Footer.tsx): add unique key prop to FooterHead, Links, and FooterSponsors components
The unique key prop is added to each of the FooterHead, Links, and FooterSponsors components. This is necessary to avoid the "Each child in a list should have a unique 'key' prop" warning in the console. The uuid4 function from the uuid library is used to generate a unique key for each component.