test(NewsPreview.test.tsx): add test for selectable prop
This commit adds a new test to the NewsPreview component that checks if the selectable prop is set to true, the component renders a checkbox. This test ensures that the selectable prop is working as expected and that the component is rendering the checkbox when it is set to true.