feat(stock-center): add shopping cart page and strain catalog page
The shopping cart page is added to display the items in the cart. The strain catalog page is added to display a list of strains with the ability to filter and search. The strain catalog page uses the useConfigureStrainCatalogSearchGraphql hook to configure the search parameters and the useQuery hook to fetch the data. The CatalogTableDisplay component is used to display the data in a table format. The useIntersectionObserver hook is used to implement infinite scrolling. The FilterDropdown and SearchBox components are used to filter and search the data.