feat(state.ts): add initial state and atoms for shopping cart
This commit adds an initial state for the shopping cart, which includes an array of added items, a boolean indicating if the cart is full, and constants for the cart limit and storage keys. It also adds atoms for the added items and isFull state, which can be used to update the state of the shopping cart. Finally, it adds a removeCartItemAtom, which is currently empty and can be used to remove items from the cart.