feat(state.ts): add cartFullAtom to check if cart is full
The cartFullAtom is added to check if the cart is full. It is a boolean atom that returns true if the total number of items in the cart is equal to the maximum number of items allowed, which is defined as MAX_ITEMS. This atom can be used to disable the "Add to Cart" button when the cart is full.