test(ShoppingCartWithItems.test.tsx): update testItems array
The testItems array has been updated to include more items with different fees. The ShoppingCartWithItems component now receives an array of items as a prop and calculates the total fee of all items in the cart. The total fee is then displayed in the component. This improves the user experience by providing a clear indication of the total cost of all items in the cart.