test(ShoppingCartTotalCard): move test for displaying order total
The test for displaying the total price was moved to ShoppingCartTotalCart.test.tsx to simplify testing. The test uses the test data from testData.tsx to render the component and check that the total is displayed correctly. The test for deleting items in cart has been temporarily marked as todo until the function is implemented.