feat(state.ts): add strainItemsAtom, maxItemsAtom, isFullAtom, and deleteItemAtom
The state has been renamed from "State" to "Cart" to improve clarity. The new atoms added are strainItemsAtom, maxItemsAtom, isFullAtom, and deleteItemAtom. strainItemsAtom and maxItemsAtom are used to get the strainItems and maxItems properties of the cart respectively. isFullAtom is used to check if the cart is full. deleteItemAtom is used to delete an item from the cart. These new atoms improve the functionality of the cart state.