feat(ShoppingCartItem.tsx): add delete button to remove item from cart
The ShoppingCartItem component now accepts a deleteItem function as a prop, which is called when the delete button is clicked. The delete button is an IconButton with a ClearIcon, and it is displayed on the right side of the CardHeader. When clicked, the deleteItem function is called with the id of the item to be removed from the cart.