refactor(ui-dsc): remove unused imports and variables, add type annotations
This commit removes unused imports and variables from the CartIcon component and adds type annotations to the CartIconProperties interface. It also removes the useEffect hook that listens for updates to local storage in other tabs, as it is no longer needed. In the ShoppingCartHeader component, a comment is added to describe its purpose. In the ShoppingCartItem component, a comment is added to describe the deleteItem callback. In the ShoppingCartTotalCard component, a comment is added to describe the items property.