feat(ui-dsc): add ShoppingCartListV2 component
The ShoppingCartListV2 component is added to the ui-dsc package. This component is responsible for rendering the shopping cart page when there are items in the cart. It displays a list of cart items and provides a callback function for deleting items. The component also includes a total card and a continue shopping card.